Mathlib and Tier 0 update using 2013 headers
Some missing functions have been added/ignored, please report any issue

+++ src/tier0/platform.cpp
@@ -68,7 +68,7 @@ double Plat_FloatTime()
 #endif
 }
 
-unsigned long Plat_MSTime()
+unsigned int Plat_MSTime()
 {
 	InitTime();
